+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- `messaging.WebpushConfig` class now supports configuring additional
+  FCM options for the features supported by the web SDK. A new
+  `messaging.WebpushFcmOptions` type has been introduced for this
+  purpose.
 - `messaging.Aps` class now supports configuring a critical sound. A new
   `messaging.CriticalSound` class has been introduced for this purpose.
 - [changed] Dropped support for Python 3.3.
@@ -24,7 +28,7 @@
   receiving realtime update events from the Firebase Database.
 - [added] The `db.reference()` method now optionally takes a `url`
   parameter. This can be used to access multiple Firebase Databases
-  in the same project more easily. 
+  in the same project more easily.
 - [added] The `messaging.WebpushNotification` type now supports
   additional parameters.